Fixes #1101 - NoSuchAuthorityCodeException in AbstractFeatureCreator.createGeometryFrom after fresh server restart
This issue was caused by 3rd party library geolatte-geom which is used in hibernate spatial. This lib does not load/cache some CRS (e.g. 4979) during the initialization but if you insert a feature with 4979. this EPSG would be cached in geolatte-geom. After a restart, this cache would be initialized but without 4979.

The solution for the SOS is to register all EPSG codes from the settings (supported and default staorage/response EPSG codes) during the initialization of the SOS .