Resolved a minor issue in the README.MD file
In this commit, I addressed a small issue in the README.MD file. There was a link present that wasn't embedded correctly, resulting in a non-functional link. To improve the documentation and user experience, I fixed this issue by properly embedding the link. Now, users will be able to access the linked resource seamlessly. This enhancement aims to make the README more informative and user-friendly.